split rdm to separate file
